==> 7.Dtw_wind_relationship.Rmd <==
---
title: "Dtw_wind_relationship"
author: "Camila Calderon"
date: "2025-01-13"
output: html_document
---

```{r}
pacman::p_load(dplyr, lubridate, ggplot2, data.table, plyr, ggpubr, emmeans, tidyverse, lme4, patchwork)
```

### Load dtw and wind data
```{r}
# read wind data
windall_df <- read.csv(file="/Users/ccalderon/ownCloud/PhDLife/P.hastatus/Thesis/Paper3/data/allpaths_winddata.csv")
# read dtw data
distsim_all <- read.csv(file="~/ownCloud/PhDLife/P.hastatus/Thesis/Paper3/data/distsimtable_commutes.csv")
```


### Difference in pairwise wind speed and pairwise DTW within individuals

```{r}
#convert to radians
windall_df$wind_dire_rad <- windall_df$wind_dire*(pi/180)

# calculate mean by ids
meanwind_id <- windall_df %>%
  mutate(date=date(timestamp))%>%
  group_by(ID, tag.local.identifier, date, direc)%>%
  dplyr::summarise(meanwind=mean(wind.speed),varwind=var(wind.speed), meanwd=mean(wind_dire_rad),varwd=var(wind_dire_rad))

# split by direction
meanwind_id_ls <- split(meanwind_id, meanwind_id$direc)

# create all pairwise combinations of dataframes for INBOUND commutes
pairs_windin<- expand.grid(meanwind_id_ls[[1]]$ID, meanwind_id_ls[[1]]$ID)
pairs_windin <- subset(pairs_windin, Var1 != Var2)

# rename columns in pairs_windout for clarity
names(pairs_windin) <- c("Pair1", "Pair2")

# merge pairs_windin with the original data twice, for ID1 and ID2
pairs_windin <- merge(pairs_windin, meanwind_id_ls[[1]], by.x = "Pair1", by.y = "ID", suffixes = c("_Pair1", ""))
pairs_windin <- merge(pairs_windin, meanwind_id_ls[[1]], by.x = "Pair2", by.y = "ID", suffixes = c("_Pair1", "_Pair2"))

# Calculate the absolute difference in meanwind for each pair
pairs_windin$windspDf <- abs(pairs_windin$meanwind_Pair1 - pairs_windin$meanwind_Pair2)
pairs_windin$windspVar <- abs(pairs_windin$varwind_Pair1 - pairs_windin$varwind_Pair2)
pairs_windin$meanwd <- abs(pairs_windin$meanwd_Pair1 - pairs_windin$meanwd_Pair2)
pairs_windin$varwd <- abs(pairs_windin$varwd_Pair1 - pairs_windin$varwd_Pair2)

colnames(pairs_windin)[5] <- "direc"


# Create all pairwise combinations of dataframes for OUTBOUND commutes
pairs_windout<- expand.grid(meanwind_id_ls[[2]]$ID, meanwind_id_ls[[2]]$ID)
pairs_windout <- subset(pairs_windout, Var1 != Var2)

# Rename columns in pairs_windout for clarity
names(pairs_windout) <- c("Pair1", "Pair2")

# Merge pairs_windout with the original data twice, for ID1 and ID2
pairs_windout <- merge(pairs_windout, meanwind_id_ls[[2]], by.x = "Pair1", by.y = "ID", suffixes = c("_Pair1", ""))
pairs_windout <- merge(pairs_windout, meanwind_id_ls[[2]], by.x = "Pair2", by.y = "ID", suffixes = c("_Pair1", "_Pair2"))

# Calculate the absolute difference in meanwind for each pair
pairs_windout$windspDf <- abs(pairs_windout$meanwind_Pair1 - pairs_windout$meanwind_Pair2)
pairs_windout$windspVar <- abs(pairs_windout$varwind_Pair1 - pairs_windout$varwind_Pair2)
pairs_windout$meanwd <- abs(pairs_windout$meanwd_Pair1 - pairs_windout$meanwd_Pair2)
pairs_windout$varwd <- abs(pairs_windout$varwd_Pair1 - pairs_windout$varwd_Pair2)
colnames(pairs_windout)[5] <- "direc"

# Merge the two data frames
pairs_wind <- rbind(pairs_windout,pairs_windin)

# Merge distsim_all with pairs_wind on Pair1 and Pair2
distsim_all <- merge(distsim_all, pairs_wind[, c("Pair1", "Pair2","direc", "windspDf", "windspVar", "meanwd","varwd")], 
                     by = c("Pair1", "Pair2", "direc"), all.x = TRUE)

### calculate crosswind and wind support
meanws <- windall_df %>%
    mutate(date=date(timestamp))%>%
    filter(date<"2022-02-28")%>%
    group_by(direc)%>%
    dplyr::summarise(meancw=mean(cw, na.rm=TRUE), meanws=mean(ws, na.rm=TRUE), sdws=sd(ws, na.rm=TRUE), varcw=var(cw, na.rm=TRUE), varws=var(ws, na.rm=TRUE))

# calculate mean crosswind by ids
meancw_ws <- windall_df %>%
  mutate(date=date(timestamp))%>%
  group_by(ID, tag.local.identifier, date, direc)%>%
  dplyr::summarise(meancw=mean(cw), meanws=mean(ws), varcw=var(cw), varws=var(ws))%>%
  filter(date<"2022-02-28")


meancw_ws_id_ls <- split(meancw_ws, meancw_ws$direc)

# create all pairwise combinations of dataframes for INBOUND commutes
pairs_cwws_in<- expand.grid(meancw_ws_id_ls[[1]]$ID, meancw_ws_id_ls[[1]]$ID)
pairs_cwws_in <- subset(pairs_cwws_in, Var1 != Var2)

# rename columns in pairs_windout for clarity
names(pairs_cwws_in) <- c("Pair1", "Pair2")

# merge pairs_windout with the original data twice, for ID1 and ID2
pairs_cwws_in <- merge(pairs_cwws_in, meancw_ws_id_ls[[1]], by.x = "Pair1", by.y = "ID", suffixes = c("_Pair1", ""))
pairs_cwws_in <- merge(pairs_cwws_in, meancw_ws_id_ls[[1]], by.x = "Pair2", by.y = "ID", suffixes = c("_Pair1", "_Pair2"))

# calculate the absolute difference in meanwind for each pair
pairs_cwws_in$difCW <- abs(pairs_cwws_in$meancw_Pair1 - pairs_cwws_in$meancw_Pair2)
pairs_cwws_in$difWS <- abs(pairs_cwws_in$meanws_Pair1 - pairs_cwws_in$meanws_Pair2)
pairs_cwws_in$varCW <- abs(pairs_cwws_in$varcw_Pair1 - pairs_cwws_in$varcw_Pair2)
pairs_cwws_in$varWS <- abs(pairs_cwws_in$varws_Pair1 - pairs_cwws_in$varws_Pair2)
colnames(pairs_cwws_in)[5] <- "direc"

# create all pairwise combinations of dataframes for OUTBOUND commutes
pairs_cwws_out<- expand.grid(meancw_ws_id_ls[[2]]$ID, meancw_ws_id_ls[[2]]$ID)
pairs_cwws_out <- subset(pairs_cwws_out, Var1 != Var2)

# rename columns in pairs_windout for clarity
names(pairs_cwws_out) <- c("Pair1", "Pair2")

# merge pairs_windout with the original data twice, for ID1 and ID2
pairs_cwws_out <- merge(pairs_cwws_out, meancw_ws_id_ls[[2]], by.x = "Pair1", by.y = "ID", suffixes = c("_Pair1", ""))
pairs_cwws_out <- merge(pairs_cwws_out, meancw_ws_id_ls[[2]], by.x = "Pair2", by.y = "ID", suffixes = c("_Pair1", "_Pair2"))

# calculate the absolute difference in meanwind for each pair
pairs_cwws_out$difCW <- abs(pairs_cwws_out$meancw_Pair1 - pairs_cwws_out$meancw_Pair2)
pairs_cwws_out$difWS <- abs(pairs_cwws_out$meanws_Pair1 - pairs_cwws_out$meanws_Pair2)
pairs_cwws_out$varCW <- abs(pairs_cwws_out$varcw_Pair1 - pairs_cwws_out$varcw_Pair2)
pairs_cwws_out$varWS <- abs(pairs_cwws_out$varws_Pair1 - pairs_cwws_out$varws_Pair2)
colnames(pairs_cwws_out)[5] <- "direc"

# merge the two data frames
pairs_cwws_inout <- rbind(pairs_cwws_out,pairs_cwws_in)

# merge distsim_all with pairs_wind on Pair1 and Pair2
distsim_all <- merge(distsim_all, pairs_cwws_inout[, c("Pair1", "Pair2","direc", "difCW","difWS", "varCW","varWS")], 
                     by = c("Pair1", "Pair2", "direc"), all.x = TRUE)


distsim_all_sameid <- distsim_all %>%
  filter(sameid=="yes")
```
### Plot wind speed against DTW distance within individuals all tracking periods
```{r}
# plot mean wind speed
winddif_dist <- ggplot(aes(x=windspDf, y=log(Distance)),data=distsim_all_sameid)+
  geom_point()+
  theme_classic()+
  ylab("Pairwise distance within \n individual commutes")+
  xlab("Mean wind speed \n difference (m/s)")+
 theme(legend.position = "none", 
        axis.text.x = element_text(size = 14),
        axis.text.y = element_text(size = 14),
        axis.title.x = element_text(size = 18),
        axis.title.y = element_text(size = 18), 
        plot.title= element_text(size = 20))+
  facet_wrap(.~direc)+
  geom_smooth(method="lm")

#variance in wind speed
winddif_var <- ggplot(aes(x=windspVar, y=log(Distance)),data=distsim_all_sameid)+
  geom_point()+
  theme_classic()+
  ylab("Pairwise DTW distance of \n commutes within ids")+
  xlab("Wind speed variance \n difference (m/s)")+
  theme(legend.position = "none", 
        axis.text.x = element_text(size = 14),
        axis.text.y = element_text(size = 14),
        axis.title.x = element_text(size = 18),
        axis.title.y = element_text(size = 18), 
        plot.title= element_text(size = 20))+
  facet_wrap(.~direc)+
  geom_smooth(method="lm")
```
### Plot crosswind and DTW distance within individuals all tracking periods
```{r}
# plot mean cross wind 
cwdiff_dist <-ggplot(aes(x=difCW, y=log(Distance)),data=distsim_all_sameid)+
  geom_point()+
  theme_classic()+
  ylab("Pairwise DTW distance of \n commutes within ids")+
  xlab("Mean crosswind \n difference (m/s)")+
 theme(legend.position = "none", 
        axis.text.x = element_text(size = 14),
        axis.text.y = element_text(size = 14),
        axis.title.x = element_text(size = 18),
        axis.title.y = element_text(size = 18), 
        plot.title= element_text(size = 20))+
  facet_wrap(.~direc)+
  geom_smooth(method="lm")

# plot variance cross wind 
cwvar_dist <-ggplot(aes(x=varCW, y=log(Distance)),data=distsim_all_sameid)+
  geom_point()+
  theme_classic()+
  ylab("Pairwise DTW distance of \n commutes within ids")+
  xlab("Crosswind variance \n difference (m/s)")+
 theme(legend.position = "none", 
        axis.text.x = element_text(size = 14),
        axis.text.y = element_text(size = 14),
        axis.title.x = element_text(size = 18),
        axis.title.y = element_text(size = 18), 
        plot.title= element_text(size = 20))+
  facet_wrap(.~direc)+
  geom_smooth(method="lm")
```
### Plot wind suppport against DTW distance within individuals all tracking periods
```{r}
# plot wind support
wsdiff_dist <- ggplot(aes(x=difWS, y=log(Distance)),data=distsim_all_sameid)+
  geom_point()+
  theme_classic()+
  ylab("Pairwise DTW distance of \n commutes within ids")+
  xlab("Mean wind support \n difference (m/s)")+
  theme(legend.position = "none", 
        axis.text.x = element_text(size = 14),
        axis.text.y = element_text(size = 14),
        axis.title.x = element_text(size = 18),
        axis.title.y = element_text(size = 18), 
        plot.title= element_text(size = 20))+
  facet_wrap(.~direc)+
  geom_smooth(method="lm")
  
# plot variance in wind support
wsvar_dist <-ggplot(aes(x=varWS, y=log(Distance)),data=distsim_all_sameid)+
  geom_point()+
  theme_classic()+
  ylab("Pairwise DTW distance of \n commutes within ids")+
  xlab("Wind support variance \n difference (m/s)")+
  theme(legend.position = "none", 
        axis.text.x = element_text(size = 14),
        axis.text.y = element_text(size = 14),
        axis.title.x = element_text(size = 18),
        axis.title.y = element_text(size = 18), 
        plot.title= element_text(size = 20))+
  facet_wrap(.~direc)+
  geom_smooth(method="lm")

winddif_dist 
```

### Plot of wind variables together to see possible relationship

It looks like there might be an influence of wind support and crosswind to DTW distance por all individuals
```{r}
# mean plot wind variables
wsdiff_dist / cwvar_dist + plot_layout(axes="collect_y") + plot_annotation(tag_levels = "A")

ggsave(file="/Users/ccalderon/ownCloud/PhDLife/P.hastatus/Thesis/Paper3/analysis/figures/MeanVar_WS_CW_DTW.pdf", width=7, height=7)

#variance plot mean variable
wsvar_dist/cwvar_dist  + plot_layout(axes="collect_y")

ggsave(file="/Users/ccalderon/ownCloud/PhDLife/P.hastatus/Thesis/Paper3/analysis/figures/Var_WS_CW_DTW.pdf", width=7, height=12)

# variance in crosswind can have an affect on both if they change more shape on the inbound commutes when they cross the ocean

(wsdiff_dist + cwvar_dist)/ combined_plot 

```

### Test DTW pairwise distance relationship with wind support and crosswind
```{r}
#Test with a linear regression
cor_matrix <- cor(distsim_all_sameid[, c("windspDf", "difCW", "difWS", "windspVar", "varCW", "varWS")], use = "complete.obs")
print(cor_matrix)

names(distsim_all_sameid)
 
### Check this better
model_path_wsp_mean <- lmer(log(Distance)~windspDf +difCW +direc + (1|dyad), data=distsim_all_sameid)

summary(model_path_wsp_mean)
car::Anova(model_path_wsp_mean)
plot(model_path_wsp_mean)

model_path_ws <- lmer(log(Distance)~difCW + difWS+direc + (1|dyad), data=distsim_all_sameid)

summary(model_path_ws)
car::Anova(model_path_ws)
plot(model_path_ws)

# Calculate marginal means for 'direc'
em_direc <- emmeans(model_path_ws, ~ direc)
pairwise_em_direc <- pairs(em_direc)

model_path_ws_var <- lmer(log(Distance)~ windspVar+varCW+varWS+direc+ (1|dyad), data=distsim_all_sameid)

summary(model_path_ws_var)
car::Anova(model_path_ws_var)
plot(model_path_ws_var)  

#Best fit
AIC(model_path_wsp_mean,model_path_ws,model_path_ws_var) # best model fit is the wind support and crosswind mean difference

```

### Calculate the predicted values for wind support and plot
```{r}
# Generate prediction data
new_data <- expand.grid(
  difWS = seq(min(distsim_all_sameid$difWS, na.rm=TRUE), max(distsim_all_sameid$difWS, na.rm=TRUE), length.out = 100),
  difCW = mean(distsim_all_sameid$difCW, na.rm=TRUE),  # Hold difWS constant
  direc = unique(distsim_all_sameid$direc)  # Include both "outbound" and "inbound"
)

new_data$predicted_log_distance <- predict(model_path_ws, newdata = new_data, re.form = NA)
new_data$predicted_se <- predict(model_path_ws, newdata = new_data, re.form = NA, se.fit = TRUE)$se.fit
new_data$lower <- new_data$predicted_log_distance - 1.96 * new_data$predicted_se
new_data$upper <- new_data$predicted_log_distance + 1.96 * new_data$predicted_se


# Plot predictions
# Add observed values of log(Distance)
distsim_all_sameid$log_distance <- log(distsim_all_sameid$Distance)
#distsim_all_sameid$direc_f <- factor(distsim_all_sameid$direc, levels = c("outbound","inbound"))

# Plot predictions with confidence intervals and observed data
ws_dist <- ggplot() +
  geom_ribbon(data = new_data, aes(x = difWS, ymin = lower, ymax = upper), alpha = 0.2) +
  geom_line(data = new_data, aes(x = difWS, y = predicted_log_distance), color="#619CFF", size = 1) +
  geom_point(data = distsim_all_sameid, aes(x = difWS, y = log_distance), alpha = 0.6) +
  scale_y_continuous(breaks = c(9, 10, 11), labels = c("8100", "22000", "59800"))+
  labs(
    x = "Mean wind support (m/s)",
    y = "Predicted log pairwise DTW distance of \n commutes within ids "
  ) +
  guides(fill="none", color="none")+
  theme(legend.position = "none", 
        axis.text.x = element_text(size = 14),
        axis.text.y = element_text(size = 14),
        axis.title.x = element_text(size = 18),
        axis.title.y = element_text(size = 18), 
        plot.title= element_text(size = 20))+
  theme_classic()+
  facet_wrap(.~direc,scales = "free")
``` 
### Calculate the predicted values for cross wind and plot
```{r}
# Generate prediction data
new_data_2 <- expand.grid(
  varCW = seq(min(distsim_all_sameid$varCW, na.rm=TRUE), max(distsim_all_sameid$varCW, na.rm=TRUE), length.out = 100),
  varWS = mean(distsim_all_sameid$varWS, na.rm=TRUE),
  windspVar = mean(distsim_all_sameid$windspVar, na.rm=TRUE),# Hold difWS constant
  direc = unique(distsim_all_sameid$direc)  # Include both "outbound" and "inbound"
)

new_data_2$predicted_log_distance <- predict(model_path_ws_var, newdata = new_data_2, re.form = NA)
new_data_2$predicted_se <- predict(model_path_ws_var, newdata = new_data_2, re.form = NA, se.fit = TRUE)$se.fit
new_data_2$lower <- new_data_2$predicted_log_distance - 1.96 * new_data_2$predicted_se
new_data_2$upper <- new_data_2$predicted_log_distance + 1.96 * new_data_2$predicted_se

# Add observed values of log(Distance)
distsim_all_sameid$log_distance_2 <- log(distsim_all_sameid$Distance)

# Plot predictions with confidence intervals and observed data
cw_dist <- ggplot() +
  geom_ribbon(data = new_data_2, aes(x = varCW, ymin = lower, ymax = upper), alpha = 0.2) +
  geom_line(data = new_data_2, aes(x = varCW, y = predicted_log_distance), color="#619CFF", size = 1) +
  geom_point(data = distsim_all_sameid, aes(x = varCW, y = log_distance_2), alpha = 0.6) +
  scale_y_continuous(breaks = c(9, 10, 11), labels = c("8100", "22000", "59800"))+
  labs(
    x = "Crosswind variance \n difference (m/s)",
    y = "Predicted log pairwise DTW distance of \n commutes within ids "
  ) +
  guides(fill="none", color="none")+
  theme(legend.position = "none", 
        axis.text.x = element_text(size = 14),
        axis.text.y = element_text(size = 14),
        axis.title.x = element_text(size = 18),
        axis.title.y = element_text(size = 18), 
        plot.title= element_text(size = 20))+
  theme_classic()+
  facet_wrap(.~direc,scales = "free")
```

### Plot Figure of wind variables and DTW distance similarity
```{r}
# plot
ws_dist / cw_dist + plot_layout(axes="collect_y") + plot_annotation(tag_levels = "A")

ggsave(file="/Users/ccalderon/ownCloud/PhDLife/P.hastatus/Thesis/Paper3/analysis/figures/Fig2.pdf", width=8, height=7)
```
