==> 8.Airspeed_crosswind_commutes.Rmd <==
---
title: "tracks_airspeed_crosswind"
author: "Camila Calderon"
date: "2025-01-13"
output: html_document
---


```{r setup, include=FALSE}
pacman::p_load(lubridate,ggplot2,tidyverse,mapview, circular, ggridges, ggmap, ggnewscale, ggspatial)
```


### Load data
```{r}
# read wind data
windall_df <- read.csv(file="/Users/ccalderon/ownCloud/PhDLife/P.hastatus/Thesis/Paper3/data/allpaths_winddata.csv")
```

### Airspeed outbound and inbound commute

airspeed (v a ) is the speed of the bird relative to the wind

```{r}
windall_out_df <- windall_df %>%
  filter(direc=="outbound")

windall_in_df <- windall_df %>%
  filter(direc=="inbound")

p_air_gs_out <- 
  windall_out_df %>%
  filter((timestamp)<"2022-02-14")%>%
  ggplot(aes(x=ground.speed, y=airspeed))+
  geom_point(aes(color=ws))+
  viridis::scale_color_viridis()+
  theme_classic()+
  xlab(expression(paste("Ground speed (ms"^"-1)")))+
  ylab(expression(paste("Airspeed (ms"^"-1)")))+
  ggtitle("Outbound commutes")+
  theme(legend.position = "bottom")

p_air_gs_in <- 
  windall_in_df %>%
  filter(date(timestamp)<"2022-02-14")%>%
  ggplot(aes(x=ground.speed, y=airspeed))+
  geom_point(aes(color=ws))+
  viridis::scale_color_viridis()+
  theme_classic()+
  xlab(expression(paste("Ground speed (ms"^"-1)")))+
  ylab(expression(paste("Airspeed (ms"^"-1)")))+
  ggtitle("Inbound commutes")+
  theme(legend.position = "bottom")

p_air_gs_out + p_air_gs_in

```

### Plot tracks in relation with airspeed

```{r}
air_out <- windall_out_df %>%
  filter(date(timestamp)<"2022-02-14")%>%
ggplot(aes(x=location.long, y=location.lat))+
  geom_point(aes(color=airspeed))+
  viridis::scale_color_viridis()+
  theme_classic()+
  labs(
    title = "Tracks of outbound commutes",
    x = "Longitude",
    y = "Latitute",
    color = "Airspeed (m/s)",
    tag = "A"
  )+
  theme(plot.tag = element_text(size = 22))+
  #facet_wrap(.~date(timestamp)) 
  facet_wrap(.~tag.local.identifier) 

air_in <- windall_in_df %>%
  filter(date(timestamp)<"2022-02-14")%>%
ggplot(aes(x=location.long, y=location.lat))+
  geom_point(aes(color=airspeed))+
  viridis::scale_color_viridis()+
  theme_classic()+
  labs(
    title = "Tracks of inbound commutes",
    x = "Longitude",
    y = "Latitute",
    color = "Airspeed (m/s)",
    tag = "B"
  )+
  theme(plot.tag = element_text(size = 22))+
  facet_wrap(.~tag.local.identifier)


# add date to dataframe
windall_out_df$date <- date(windall_out_df$timestamp)

# Plot density by date
den_plot_out <- windall_out_df %>%
  filter(as.Date(timestamp) < as.Date("2022-02-14")) %>%
  ggplot(aes(x = airspeed, y = as.factor(date), fill = ..x..)) +
  geom_density_ridges_gradient(
    scale = 1.5,
    quantile_lines = TRUE,
    quantiles = 2
  ) +
  scale_fill_viridis_c(option = "C") +
  theme_classic() +
  labs(
    title = "Density plots of airspeed of outbound commutes",
    x = "Airspeed (m/s)",
    y = "Date",
    fill = "Airspeed (m/s)",
    tag=""
  )+
  facet_wrap(.~tag.local.identifier)

# add date to dataframe
windall_in_df$date <- date(windall_in_df$timestamp)

# Plot density by date
den_plot_in <-windall_in_df %>%
  filter(as.Date(timestamp) < as.Date("2022-02-14")) %>%
  ggplot(aes(x = airspeed, y = as.factor(date), fill = ..x..)) +
  geom_density_ridges_gradient(
    scale = 1.5,
    quantile_lines = TRUE,
    quantiles = 2
  ) +
  scale_fill_viridis_c(option = "C") +
  theme_classic() +
  labs(
    title = "Density plots of airspeed of inbound commutes",
    x = "Airspeed (m/s)",
    y = "Date",
    fill = "Airspeed (m/s)",
    tag=""
  )+
  facet_wrap(.~tag.local.identifier)

(air_out/ den_plot_out) 
ggsave(file="/Users/ccalderon/ownCloud/PhDLife/P.hastatus/Thesis/Paper3/analysis/figures/airspeed_id_out.pdf", width = 15, height = 20)

(air_in/den_plot_in)
ggsave(file="/Users/ccalderon/ownCloud/PhDLife/P.hastatus/Thesis/Paper3/analysis/figures/airspeed_id_in.pdf", width = 15, height = 20)

```

### Plot tracks in relation with wind support

```{r}
plot_ws_out <- windall_out_df %>%
  filter(date(timestamp)<"2022-02-14")%>%
ggplot(aes(x=location.long, y=location.lat))+
  geom_point(aes(color=ws))+
  viridis::scale_color_viridis()+
  theme_classic()+
  labs(
    title = "Tracks of outbound commutes",
    x = "Longitude",
    y = "Latitute",
    color = "Wind support"
  )+
  facet_wrap(.~tag.local.identifier)

den_plotws_out <-windall_out_df %>%
  filter(as.Date(timestamp) < as.Date("2022-02-14")) %>%
  ggplot(aes(x = ws, y = as.factor(date), fill = ..x..)) +
  geom_density_ridges_gradient(
    scale = 1.5,
    quantile_lines = TRUE,
    quantiles = 2
  ) +
  scale_fill_viridis_c(option = "C") +
  theme_classic() +
  labs(
    title = "Density plots of wind support of outbound commutes",
    x = "Wind support",
    y = "Date",
    fill = "Wind support"
  )+
  facet_wrap(.~tag.local.identifier)


plot_ws_in <- windall_in_df %>%
  filter(date(timestamp)<"2022-02-14")%>%
ggplot(aes(x=location.long, y=location.lat))+
  geom_point(aes(color=ws))+
  viridis::scale_color_viridis()+
  theme_classic()+
  labs(
    title = "Tracks of inbound commutes",
    x = "Longitude",
    y = "Latitute",
    color = "Wind support"
  )+
  facet_wrap(.~tag.local.identifier)

den_plotws_in <-windall_in_df %>%
  filter(as.Date(timestamp) < as.Date("2022-02-14")) %>%
  ggplot(aes(x = ws, y = as.factor(date), fill = ..x..)) +
  geom_density_ridges_gradient(
    scale = 1.5,
    quantile_lines = TRUE,
    quantiles = 2
  ) +
  scale_fill_viridis_c(option = "C") +
  theme_classic() +
  labs(
    title = "Density plots of wind support of inbound commutes",
    x = "Wind support",
    y = "Date",
    fill = "Wind support"
  )+
  facet_wrap(.~tag.local.identifier)

plot_ws_out/den_plotws_out

ggsave(file="/Users/ccalderon/ownCloud/PhDLife/P.hastatus/Thesis/Paper3/analysis/figures/windsupport_id_out.pdf", width = 15, height = 20)

plot_ws_in/den_plotws_in

ggsave(file="/Users/ccalderon/ownCloud/PhDLife/P.hastatus/Thesis/Paper3/analysis/figures/windsupport_id_in.pdf", width = 15, height = 20)

```

### Plot tracks in relation with cross wind
```{r}
cw_out <-windall_out_df %>%
  filter(date(timestamp)<"2022-02-14")%>%
ggplot(aes(x=location.long, y=location.lat))+
  geom_point(aes(color=abs(cw)))+
  viridis::scale_color_viridis()+
  theme_classic()+
  labs(
    title = "Tracks of outbound commutes",
    x = "Longitude",
    y = "Latitute",
    color = "Crosswind (m/s)",
    tag = "A"
  )+
  theme(plot.tag = element_text(size=22))+
  facet_wrap(.~tag.local.identifier)

den_plotcw_out <-windall_out_df %>%
  filter(as.Date(timestamp) < as.Date("2022-02-14")) %>%
  ggplot(aes(x = abs(cw), y = as.factor(date), fill = ..x..)) +
  geom_density_ridges_gradient(
    scale = 1.5,
    quantile_lines = TRUE,
    quantiles = 2
  ) +
  scale_fill_viridis_c(option = "C") +
  theme_classic() +
  labs(
    title = "Density plots of crosswind of outbound commutes",
    x = "Crosswind (m/s)",
    y = "Date",
    fill = "Crosswind (m/s)",
    tag = ""
  )+
  facet_wrap(.~tag.local.identifier)

cw_in <-windall_in_df %>%
  filter(date(timestamp)<"2022-02-14")%>%
ggplot(aes(x=location.long, y=location.lat))+
  geom_point(aes(color=abs(cw)))+
  viridis::scale_color_viridis()+
  theme_classic()+
  labs(
    title = "Tracks of inbound commutes",
    x = "Longitude",
    y = "Latitute",
    color = "Crosswind (m/s)",
    tag = "B"
  )+
  theme(plot.tag = element_text(size=22))+
  facet_wrap(.~tag.local.identifier)


den_plotcw_in <-windall_in_df %>%
  filter(as.Date(timestamp) < as.Date("2022-02-14")) %>%
  ggplot(aes(x = abs(cw), y = as.factor(date), fill = ..x..)) +
  geom_density_ridges_gradient(
    scale = 1.5,
    quantile_lines = TRUE,
    quantiles = 2
  ) +
  scale_fill_viridis_c(option = "C") +
  theme_classic() +
  labs(
    title = "Density plots of crosswind of inbound commutes",
    x = "Crosswind (m/s)",
    y = "Date",
    fill = "Crosswind (m/s)",
    tag = ""
  )+
  facet_wrap(.~tag.local.identifier)


cw_out / den_plotcw_out 
ggsave(file="/Users/ccalderon/ownCloud/PhDLife/P.hastatus/Thesis/Paper3/analysis/figures/crosswind_id_out.pdf", width = 15, height = 20)

cw_in / den_plotcw_in
ggsave(file="/Users/ccalderon/ownCloud/PhDLife/P.hastatus/Thesis/Paper3/analysis/figures/crosswind_id_in.pdf", width = 15, height = 20)
```

### Plot example of aispeed of one individual
```{r}
# plot using stamen map from Bocas
register_stadiamaps("") 

# create a bounding box
e <- make_bbox(location.long, location.lat, data =windall_df)

# make a data frame with the locations of the colonies
lagruta <- data.frame(colonies=c("lagruta"), location_lat=c( 9.396448), location_long=c(-82.271541))

# Define common limits for longitude and latitude
x_limits <- c(-82.50, -82.25)  # Adjust as needed
y_limits <- c(9.34, 9.44)    # Adjust as needed
y_breaks <- seq(9.36, 9.44, by = 0.03)
x_breaks <- seq(-82.50, -82.25, by = 0.07)

#Airspeed
airspeed_id1_out <- get_stadiamap(e, zoom = 12, maptype = "stamen_toner_lite") %>% ggmap()+
  coord_sf(crs = "+proj=lonlat", expand = FALSE)+
  #scale_x_continuous(breaks = c(-82.50, -82.40, -82.30))+
  geom_point(data=windall_out_df[windall_out_df$tag.local.identifier=="2D507235_G",], aes(x=location.long, y=location.lat, color=airspeed))+
  viridis::scale_color_viridis(name="Airspeed (m/s)")+
  new_scale_color() +
  geom_point(data=lagruta, aes(x=location_long, y=location_lat, shape=colonies, size=5, fill=colonies, color=colonies),color="black")+
  scale_x_continuous(limits = x_limits, breaks = x_breaks) +
  scale_y_continuous(limits = y_limits, breaks = y_breaks) +
  #annotation_scale(location = "bl", line_width = 1, height = unit(0.5, "cm"), pad_x= unit(0.5, "cm"), pad_y= unit(0.5, "cm"), text_cex = 2)+
  #annotate(geom = "text", x = -82.2, y = 9.45, label = "N=51", color = "black", size = 6) +
  theme_linedraw()+
  guides(fill= "none",
         size="none",
         color="none", 
         shape="none")+
  theme(legend.position = "bottom", 
        axis.text.x = element_text(size = 14),
        axis.text.y = element_text(size = 14),
        axis.title.x = element_text(size = 18),
        axis.title.y = element_text(size = 18), 
        plot.title= element_text(size = 20))+
 xlab("Longitude")+
  ylab("Latitude")

airspeed_id1_in <- get_stadiamap(e, zoom = 12, maptype = "stamen_toner_lite") %>% ggmap()+
  coord_sf(crs = "+proj=lonlat", expand = FALSE)+
  #scale_x_continuous(breaks = c(-82.50, -82.40, -82.30))+
  geom_point(data=windall_in_df[windall_in_df$tag.local.identifier=="2D507235_G",], aes(x=location.long, y=location.lat, color=airspeed))+
  viridis::scale_color_viridis(name="Airspeed (m/s)")+
  new_scale_color() +
  geom_point(data=lagruta, aes(x=location_long, y=location_lat, shape=colonies, size=5, fill=colonies, color=colonies), color="black")+
  scale_x_continuous(limits = x_limits, breaks = x_breaks) +
  scale_y_continuous(limits = y_limits, breaks = y_breaks) +
  annotation_scale(location = "bl", line_width = 0.8, height = unit(0.5, "cm"), pad_x= unit(0.5, "cm"), pad_y= unit(0.5, "cm"), text_cex = 1)+
  #annotate(geom = "text", x = -82.2, y = 9.45, label = "N=51", color = "black", size = 6) +
  theme_linedraw()+
  guides(fill= "none",
         size="none",
         color="none", 
         shape="none")+
  theme(legend.position = "bottom", 
        axis.text.x = element_text(size = 14),
        axis.text.y = element_text(size = 14),
        axis.title.x = element_text(size = 18),
        axis.title.y = element_text(size = 18), 
        plot.title= element_text(size = 20))+
 xlab("Longitude")+
  ylab("Latitude")


## croswind
cw_id1_out <- get_stadiamap(e, zoom = 12, maptype = "stamen_toner_lite") %>% ggmap()+
  coord_sf(crs = "+proj=lonlat", expand = FALSE)+
  #scale_x_continuous(breaks = c(-82.50, -82.40, -82.30))+
  geom_point(data=windall_out_df[windall_out_df$tag.local.identifier=="2D507235_G",], aes(x=location.long, y=location.lat, color=abs(cw)))+
  viridis::scale_color_viridis(name="Crosswind (m/s)", option = "magma")+
  new_scale_color() +
  geom_point(data=lagruta, aes(x=location_long, y=location_lat, shape=colonies, size=5, fill=colonies, color=colonies), color="black")+
  scale_x_continuous(limits = x_limits, breaks = x_breaks) +
  scale_y_continuous(limits = y_limits, breaks = y_breaks) +
  #annotation_scale(location = "bl", line_width = 1, height = unit(0.5, "cm"), pad_x= unit(0.5, "cm"), pad_y= unit(0.5, "cm"), text_cex = 2)+
  #annotate(geom = "text", x = -82.2, y = 9.45, label = "N=51", color = "black", size = 6) +
  theme_linedraw()+
  guides(fill= "none",
         size="none",
         color="none", 
         shape="none")+
  theme(legend.position = "bottom", 
        axis.text.x = element_text(size = 14),
        axis.text.y = element_text(size = 14),
        axis.title.x = element_text(size = 18),
        axis.title.y = element_text(size = 18), 
        plot.title= element_text(size = 20))+
 xlab("Longitude")+
  ylab("Latitude")


cw_id1_in <- get_stadiamap(e, zoom = 12, maptype = "stamen_toner_lite") %>% ggmap()+
  coord_sf(crs = "+proj=lonlat", expand = FALSE)+
  #scale_x_continuous(breaks = c(-82.50, -82.40, -82.30))+
  geom_point(data=windall_in_df[windall_in_df$tag.local.identifier=="2D507235_G",], aes(x=location.long, y=location.lat, color=abs(cw)))+
  viridis::scale_color_viridis(name="Crosswind (m/s)", , option = "magma")+
  new_scale_color() +
  geom_point(data=lagruta, aes(x=location_long, y=location_lat, shape=colonies, size=5, fill=colonies, color=colonies),color="black")+
  scale_x_continuous(limits = x_limits,  breaks = x_breaks) +
  scale_y_continuous(limits = y_limits, breaks = y_breaks) +
  #annotation_scale(location = "bl", line_width = 0.8, height = unit(0.5, "cm"), pad_x= unit(0.5, "cm"), pad_y= unit(0.5, "cm"), text_cex = 1)+
  #annotate(geom = "text", x = -82.2, y = 9.45, label = "N=51", color = "black", size = 6) +
  theme_linedraw()+
  guides(fill= "none",
         size="none",
         color="none", 
         shape="none")+
  theme(legend.position = "bottom", 
        axis.text.x = element_text(size = 14),
        axis.text.y = element_text(size = 14),
        axis.title.x = element_text(size = 18),
        axis.title.y = element_text(size = 18), 
        plot.title= element_text(size = 20))+
 xlab("Longitude")+
  ylab("Latitude")


# Combine plots and collect axes and guides
combined_plot <- ((airspeed_id1_out | airspeed_id1_in) / (cw_id1_out | cw_id1_in)) + plot_layout(axis_titles = "collect")+
  plot_annotation(tag_levels = "A") & 
  theme(legend.position = "bottom", plot.tag = element_text(size=18)) 
combined_plot


ggsave(file="~/ownCloud/PhDLife/P.hastatus/Thesis/Paper3/analysis/figures/airspseed_wd_id1.pdf", width=12, height = 8)
```
